requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
dnspython==2.3.0
docker==6.1.3
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.100.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
orjson==3.9.2
packaging==23.1
paho-mqtt==1.6.1
pydantic==2.0.2
pydantic-extra-types==2.0.0
pydantic-settings==2.0.1
pydantic_core==2.1.2
PyJWT==2.7.0
pymongo==4.4.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
ujson==5.8.0
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websocket-client==1.6.1
websockets==11.0.3
